requirements.txt
arrow==1.2.2
celery==5.3.4
Django==3.2.12
django-cors-headers==3.13.0
django-extensions==3.1.5
djangorestframework==3.12.4
freezegun==1.2.1
gunicorn==20.1.0
isort==5.10.1
ipython==8.1.0
jupyter==1.0.0
jupyterlab==3.3.0
mixer==7.2.2
pandas
psycopg2==2.9.3
pytest==7.0.1
pytest-django==4.5.2
pytest-mock==3.7.0
redis==4.6.0
surprise==0.1
